\name{led8x8}
\alias{led8x8}
\title{
Create a scroll text on 8x8 LED
}
\description{
Create a scroll text on 8x8 LED
}
\usage{
led8x8(txt)
}
\arguments{
  \item{txt}{string to be shown on the LED}
}
\value{
None
}
\note{
R has to be launched by sudo.
}

\section{Warning }{Not warranty} 

\examples{
led8x8("Hello World")
}
