\name{RpiGPIO-package}
\alias{RpiGPIO-package}
\alias{RpiGPIO}
\docType{package}
\title{
Demonstration of Raspberry Pi's GPIO power
}
\description{
This package provides some applications of GPIO power in Raspberry Pi by using R.
}
\details{
\tabular{ll}{
Package: \tab RpiGPIO\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-13\cr
License: \tab GNU\cr
}
The package needs to be built and then used with root access. This is, R has to be launched by sudo R.
}
\author{
MC Li

Maintainer: MC Li <arraytools@gmail.com>
}

